11 - move BeersLiked to container and change BeersLikedContainer to    extend react Component
